server/requirements.txt
flask
flask-cors
flask-jwt-extended
requests
python-dotenv
redis